requirements.txt
Flask
sqlalchemy
Flask-SQLAlchemy
argon2-cffi
click
